+++ examples/fastapi-demo/app.py
@@ -50,8 +50,8 @@ postgres = PostgresProvider(POSTGRES_URL)
 
 lock_backend = redis.lock()
 cache_backend = redis.cache()
-ratelimiter_backend = redis.ratelimiter()
-breaker_backend = postgres.circuitbreaker()
+ratelimiter_backend = redis.rate_limiter()
+breaker_backend = postgres.circuit_breaker()
 leader_backend = redis.leader_election()
 
 tasks = Tasks()